+++ jakarta-ee/flowlogix-datamodel/src/main/java/com/flowlogix/jeedao/primefaces/internal/JPAModelImpl.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.flowlogix.jeedao.primefaces.impl;
+package com.flowlogix.jeedao.primefaces.internal;
 
 import com.flowlogix.jeedao.DaoHelper;
 import com.flowlogix.jeedao.DaoHelper.QueryCriteria;
